+++ components/rating_service/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DECIMAL, TIMESTAMP
+from sqlalchemy import Column, BigInteger, TIMESTAMP, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.sql import func
 
@@ -10,5 +10,5 @@ class ProfileRating(Base):
     __tablename__ = 'profile_ratings'
 
     profile_telegram_id = Column(BigInteger, primary_key=True, index=True)
-    rating_score = Column(DECIMAL(5, 2)) # Example precision and scale for a rating score
+    rating_score = Column(Float)
     last_calculated_at = Column(TIMESTAMP, server_default=func.now())
